[Fix] Smooth scrolling delta time

+++ src/ui/button/ui_button_widget.cpp
@@ -32,7 +32,7 @@ void ekg::ui::button_widget::on_reload() {
     float text_width {f_renderer.get_text_width(ui->get_text())};
     float text_height {f_renderer.get_text_height()};
 
-    float dimension_offset {text_height / 2};
+    float dimension_offset {static_cast<float>((int32_t) (text_height / 2.0f))};
     float offset {ekg::find_min_offset(text_width, dimension_offset)};
 
     this->dimension.w = ekg::min(this->dimension.w, text_height);
@@ -48,7 +48,7 @@ void ekg::ui::button_widget::on_reload() {
     layout.process_layout_mask();
 
     auto &layout_mask {layout.get_layout_mask()};
-    this->dimension.w = this->dimension.w <= text_height ? layout_mask.w : this->dimension.w;
+    this->dimension.w = (int32_t)(this->dimension.w <= text_height ? layout_mask.w : this->dimension.w);
     this->dimension.h = ekg::min(this->dimension.h, layout_mask.h);
 }
